requirements.txt
ir-datasets>=0.5.6
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
ranx>=0.3.20
pandas
numpy
tqdm
torch>=2.0.0
psutil
accelerate>=1.1.0